requirements.txt
streamlit==1.13.0
openai==0.27.0
PyPDF2==3.0.1
Pillow==9.4.0
